Add default padding and document IE/Edge pitfalls
Add the ability to pad the right side of the input field so that
entered text doesn't collide with the clear button. The default is
20px and can be overridden.

Add a section to the bottom of the README file with a note about the
built-in clear buttons in IE/Edge and a link to more info.